ifs-kernel-updates (1:3.10.0-514-724-1ifs+deb9) UNRELEASED; urgency=medium

    * Port to linux-image 4.9.0-4.
    * Upstream version is now the actual upstream version.
    * Supported kernel version in package name and control depends.
    * Add debian/bin/rebuild_ifs_kernel_updates to assist building from
      source package.

ifs-kernel-updates (3.16.0-724-1ifs) UNRELEASED; urgency=medium

    * Import ifs-kernel-updates 3.10.0_514 724 from
      IFS RHEL73 10.6.1.0.2 upstream.
    * Version has gone backwards due to use of different
      upstream source. Previous versions were based on SLES121
      IFS.
    * Using git packaging and pristine-tar.
    * Change architecture to amd64.

+++ hfi1/aspm.h
@@ -234,7 +234,7 @@ static inline void aspm_disable_all(struct hfi1_devdata *dd)
 {
 	struct hfi1_ctxtdata *rcd;
 	unsigned long flags;
-	unsigned i;
+	u16 i;
 
 	for (i = 0; i < dd->first_user_ctxt; i++) {
 		rcd = dd->rcd[i];
@@ -253,7 +253,7 @@ static inline void aspm_enable_all(struct hfi1_devdata *dd)
 {
 	struct hfi1_ctxtdata *rcd;
 	unsigned long flags;
-	unsigned i;
+	u16 i;
 
 	aspm_enable(dd);
 
@@ -281,7 +281,7 @@ static inline void aspm_ctx_init(struct hfi1_ctxtdata *rcd)
 
 static inline void aspm_init(struct hfi1_devdata *dd)
 {
-	unsigned i;
+	u16 i;
 
 	spin_lock_init(&dd->aspm_lock);
 	dd->aspm_supported = aspm_hw_l1_supported(dd);
